src/test/java/com/abc/domain/account/CheckingAccountTest.java
package com.abc.domain.account;

import static org.hamcrest.CoreMatchers.is;
import static org.junit.Assert.assertThat;

import org.junit.Ignore;
import org.junit.Test;

import com.abc.domain.sub.money.Money;
import com.abc.domain.sub.time.DayOfYear;
import com.abc.domain.sub.time.TimePoint;

public class CheckingAccountTest {

	@Test
	public void testBalance() {
		Account account = new CheckingAccount();

		account.deposit(Money.dollars(1), TimePoint.at(2014, 3, 12, 20, 40));
		assertThat(account.balance(), is(Money.dollars(1)));
		
		account.deposit(Money.dollars(10), TimePoint.at(2014, 3, 12, 21, 40));
		assertThat(account.balance(), is(Money.dollars(11)));

		account.withdraw(Money.dollars(5), TimePoint.at(2014, 3, 12, 22, 40));
		assertThat(account.balance(), is(Money.dollars(6)));
		
		assertThat(account.entries().size(), is(3));
	}
	
	@Test(expected=InSufficientAmountException.class)
	public void testWithdrawInsuffienceAmount() throws Exception {
		Account account = new CheckingAccount();
		
		account.deposit(Money.dollars(9), TimePoint.at(2014, 3, 12, 20, 40));
		account.withdraw(Money.dollars(10), TimePoint.at(2014, 3, 12, 21, 40));
	}
	
	/**
	 * $1,000 * (10 / 365) =  $27.39 * 0.001 = $0.03
	 */
	@Test
	public void testInterestDeposit10DaysAgo() throws Exception {
		CheckingAccount account = new CheckingAccount();
		account.deposit(Money.dollars(1000), TimePoint.at(2014, 3, 2, 20, 40));
		
		account.today = DayOfYear.at(2014, 3, 12);
		assertThat(account.interest(), is(Money.dollars(0.03)));
	}
	
	/**
	 * $3,000 * (10 / 365) =  $246.57 * 0.001 = $0.25
	 * $1,000 * (10 / 365) =  $27.39 * 0.001 = $0.03
	 */
	@Test
	@Ignore("returns 0.27.. why??")
	public void testInterestDeposit30And10DaysAgo() throws Exception {
		CheckingAccount account = new CheckingAccount();
		account.deposit(Money.dollars(3000), TimePoint.at(2014, 2, 10, 20, 40));
		account.deposit(Money.dollars(1000), TimePoint.at(2014, 3, 2, 20, 40));
		
		account.today = DayOfYear.at(2014, 3, 12);
		assertThat(account.interest(), is(Money.dollars(0.28)));
	}
}
